requirements.txt
spacy==3.5.0
nltk>=3.6.0,<4.0.0
pandas>=1.2.0,<2.0.0
pdfminer.six>=20200517
python-docx>=0.8.10
docx2txt>=0.8
numpy==1.22.0